alterando pra long

+++ grails-app/controllers/com/desafio/domain/PayerController.groovy
@@ -43,7 +43,7 @@ class PayerController {
     }
 
     def show() {
-        return [payer: payerService.getPayer(params.int('id'))]
+        return [payer: payerService.getPayer(params.long('id'))]
     }
 
 }
